requirements.txt
requests
PyYAML
pytest
openapi-spec-validator